requirements.txt
flask>=3.0
rasterio>=1.3
numpy>=1.24
Pillow>=10.0
pyproj>=3.6
shapely>=2.0
fiona>=1.9
geopandas>=0.14
torch>=2.1
torchvision>=0.16
sam2>=1.0
opencv-python>=4.8
